backend/requirements.txt
flask==2.0.1
twilio==7.17.0
transformers==4.30.2
python-dotenv==0.19.0
torch==2.0.1 